Extract NodeHandlerWrapper
