+++ core/src/protocols_handler/mod.rs
@@ -18,29 +18,24 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::{
-    nodes::handled_node::{NodeHandler, NodeHandlerEndpoint, NodeHandlerEvent},
-    upgrade::{
-        self,
-        InboundUpgrade,
-        OutboundUpgrade,
-        InboundUpgradeApply,
-        OutboundUpgradeApply,
-    }
+use crate::upgrade::{
+    InboundUpgrade,
+    OutboundUpgrade,
 };
 use futures::prelude::*;
 use std::{io, time::Duration};
 use tokio_io::{AsyncRead, AsyncWrite};
-use tokio_timer::Timeout;
 
 pub use self::dummy::DummyProtocolsHandler;
 pub use self::map_in::MapInEvent;
 pub use self::map_out::MapOutEvent;
+pub use self::node_handler::{NodeHandlerWrapper, NodeHandlerWrapperBuilder};
 pub use self::select::ProtocolsHandlerSelect;
 
 mod dummy;
 mod map_in;
 mod map_out;
+mod node_handler;
 mod select;
 
 /// Handler for a set of protocols for a specific connection with a remote.
@@ -187,11 +182,7 @@ pub trait ProtocolsHandler {
     where
         Self: Sized,
     {
-        NodeHandlerWrapperBuilder {
-            handler: self,
-            in_timeout: Duration::from_secs(10),
-            out_timeout: Duration::from_secs(10),
-        }
+        NodeHandlerWrapperBuilder::new(self, Duration::from_secs(10), Duration::from_secs(10))
     }
 
     /// Builds an implementation of `NodeHandler` that handles this protocol exclusively.
@@ -280,218 +271,5 @@ impl<TConnectionUpgrade, TOutboundOpenInfo, TCustom>
             }
             ProtocolsHandlerEvent::Custom(val) => ProtocolsHandlerEvent::Custom(map(val)),
         }
-    }
-}
-
-/// Prototype for a `NodeHandlerWrapper`.
-pub struct NodeHandlerWrapperBuilder<TProtoHandler>
-where
-    TProtoHandler: ProtocolsHandler,
-{
-    /// The underlying handler.
-    handler: TProtoHandler,
-    /// Timeout for incoming substreams negotiation.
-    in_timeout: Duration,
-    /// Timeout for outgoing substreams negotiation.
-    out_timeout: Duration,
-}
-
-impl<TProtoHandler> NodeHandlerWrapperBuilder<TProtoHandler>
-where
-    TProtoHandler: ProtocolsHandler
-{
-    /// Sets the timeout to use when negotiating a protocol on an ingoing substream.
-    #[inline]
-    pub fn with_in_negotiation_timeout(mut self, timeout: Duration) -> Self {
-        self.in_timeout = timeout;
-        self
-    }
-
-    /// Sets the timeout to use when negotiating a protocol on an outgoing substream.
-    #[inline]
-    pub fn with_out_negotiation_timeout(mut self, timeout: Duration) -> Self {
-        self.out_timeout = timeout;
-        self
-    }
-
-    /// Builds the `NodeHandlerWrapper`.
-    #[inline]
-    pub fn build(self) -> NodeHandlerWrapper<TProtoHandler> {
-        NodeHandlerWrapper {
-            handler: self.handler,
-            negotiating_in: Vec::new(),
-            negotiating_out: Vec::new(),
-            in_timeout: self.in_timeout,
-            out_timeout: self.out_timeout,
-            queued_dial_upgrades: Vec::new(),
-            unique_dial_upgrade_id: 0,
-        }
-    }
-}
-
-/// Wraps around an implementation of `ProtocolsHandler`, and implements `NodeHandler`.
-// TODO: add a caching system for protocols that are supported or not
-pub struct NodeHandlerWrapper<TProtoHandler>
-where
-    TProtoHandler: ProtocolsHandler,
-{
-    /// The underlying handler.
-    handler: TProtoHandler,
-    /// Futures that upgrade incoming substreams.
-    negotiating_in:
-        Vec<Timeout<InboundUpgradeApply<TProtoHandler::Substream, TProtoHandler::InboundProtocol>>>,
-    /// Futures that upgrade outgoing substreams. The first element of the tuple is the userdata
-    /// to pass back once successfully opened.
-    negotiating_out: Vec<(
-        TProtoHandler::OutboundOpenInfo,
-        Timeout<OutboundUpgradeApply<TProtoHandler::Substream, TProtoHandler::OutboundProtocol>>,
-    )>,
-    /// Timeout for incoming substreams negotiation.
-    in_timeout: Duration,
-    /// Timeout for outgoing substreams negotiation.
-    out_timeout: Duration,
-    /// For each outbound substream request, how to upgrade it. The first element of the tuple
-    /// is the unique identifier (see `unique_dial_upgrade_id`).
-    queued_dial_upgrades: Vec<(u64, TProtoHandler::OutboundProtocol)>,
-    /// Unique identifier assigned to each queued dial upgrade.
-    unique_dial_upgrade_id: u64,
-}
-
-impl<TProtoHandler> NodeHandler for NodeHandlerWrapper<TProtoHandler>
-where
-    TProtoHandler: ProtocolsHandler,
-    <TProtoHandler::OutboundProtocol as OutboundUpgrade<<TProtoHandler as ProtocolsHandler>::Substream>>::Error: std::fmt::Debug
-{
-    type InEvent = TProtoHandler::InEvent;
-    type OutEvent = TProtoHandler::OutEvent;
-    type Substream = TProtoHandler::Substream;
-    // The first element of the tuple is the unique upgrade identifier
-    // (see `unique_dial_upgrade_id`).
-    type OutboundOpenInfo = (u64, TProtoHandler::OutboundOpenInfo);
-
-    fn inject_substream(
-        &mut self,
-        substream: Self::Substream,
-        endpoint: NodeHandlerEndpoint<Self::OutboundOpenInfo>,
-    ) {
-        match endpoint {
-            NodeHandlerEndpoint::Listener => {
-                let protocol = self.handler.listen_protocol();
-                let upgrade = upgrade::apply_inbound(substream, protocol);
-                let with_timeout = Timeout::new(upgrade, self.in_timeout);
-                self.negotiating_in.push(with_timeout);
-            }
-            NodeHandlerEndpoint::Dialer((upgrade_id, user_data)) => {
-                let pos = match self
-                    .queued_dial_upgrades
-                    .iter()
-                    .position(|(id, _)| id == &upgrade_id)
-                {
-                    Some(p) => p,
-                    None => {
-                        debug_assert!(false, "Received an upgrade with an invalid upgrade ID");
-                        return;
-                    }
-                };
-
-                let (_, proto_upgrade) = self.queued_dial_upgrades.remove(pos);
-                let upgrade = upgrade::apply_outbound(substream, proto_upgrade);
-                let with_timeout = Timeout::new(upgrade, self.out_timeout);
-                self.negotiating_out.push((user_data, with_timeout));
-            }
-        }
-    }
-
-    #[inline]
-    fn inject_inbound_closed(&mut self) {
-        self.handler.inject_inbound_closed();
-    }
-
-    fn inject_outbound_closed(&mut self, user_data: Self::OutboundOpenInfo) {
-        let pos = match self
-            .queued_dial_upgrades
-            .iter()
-            .position(|(id, _)| id == &user_data.0)
-        {
-            Some(p) => p,
-            None => {
-                debug_assert!(
-                    false,
-                    "Received an outbound closed error with an invalid upgrade ID"
-                );
-                return;
-            }
-        };
-
-        self.queued_dial_upgrades.remove(pos);
-        self.handler
-            .inject_dial_upgrade_error(user_data.1, io::ErrorKind::ConnectionReset.into());
-    }
-
-    #[inline]
-    fn inject_event(&mut self, event: Self::InEvent) {
-        self.handler.inject_event(event);
-    }
-
-    #[inline]
-    fn shutdown(&mut self) {
-        self.handler.shutdown();
-    }
-
-    fn poll(&mut self) -> Poll<Option<NodeHandlerEvent<Self::OutboundOpenInfo, Self::OutEvent>>, io::Error> {
-        // Continue negotiation of newly-opened substreams on the listening side.
-        // We remove each element from `negotiating_in` one by one and add them back if not ready.
-        for n in (0..self.negotiating_in.len()).rev() {
-            let mut in_progress = self.negotiating_in.swap_remove(n);
-            match in_progress.poll() {
-                Ok(Async::Ready(upgrade)) =>
-                    self.handler.inject_fully_negotiated_inbound(upgrade),
-                Ok(Async::NotReady) => self.negotiating_in.push(in_progress),
-                // TODO: return a diagnostic event?
-                Err(_err) => {}
-            }
-        }
-
-        // Continue negotiation of newly-opened substreams.
-        // We remove each element from `negotiating_out` one by one and add them back if not ready.
-        for n in (0..self.negotiating_out.len()).rev() {
-            let (upgr_info, mut in_progress) = self.negotiating_out.swap_remove(n);
-            match in_progress.poll() {
-                Ok(Async::Ready(upgrade)) => {
-                    self.handler.inject_fully_negotiated_outbound(upgrade, upgr_info);
-                }
-                Ok(Async::NotReady) => {
-                    self.negotiating_out.push((upgr_info, in_progress));
-                }
-                Err(err) => {
-                    let msg = format!("Error while upgrading: {:?}", err);
-                    let err = io::Error::new(io::ErrorKind::Other, msg);
-                    self.handler.inject_dial_upgrade_error(upgr_info, err);
-                }
-            }
-        }
-
-        // Poll the handler at the end so that we see the consequences of the method calls on
-        // `self.handler`.
-        match self.handler.poll()? {
-            Async::Ready(Some(ProtocolsHandlerEvent::Custom(event))) => {
-                return Ok(Async::Ready(Some(NodeHandlerEvent::Custom(event))));
-            }
-            Async::Ready(Some(ProtocolsHandlerEvent::OutboundSubstreamRequest {
-                upgrade,
-                info,
-            })) => {
-                let id = self.unique_dial_upgrade_id;
-                self.unique_dial_upgrade_id += 1;
-                self.queued_dial_upgrades.push((id, upgrade));
-                return Ok(Async::Ready(Some(
-                    NodeHandlerEvent::OutboundSubstreamRequest((id, info)),
-                )));
-            }
-            Async::Ready(None) => return Ok(Async::Ready(None)),
-            Async::NotReady => (),
-        };
-
-        Ok(Async::NotReady)
     }
 }
